requirements.txt
appnope==0.1.2
APScheduler==3.6.3
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
debug==0.3.2
debugger==1.3
debugpy==1.5.1
decorator==5.1.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ipdb==0.13.9
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.0
keyboa==3.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
os-service-types==1.7.0
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
pyrsistent==0.18.0
pyTelegramBotAPI==4.1.1
python-decouple==3.5
python-dotenv==0.19.1
python-telegram-bot==13.7
python-telegram-bot-raw==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
see==1.4.1
six==1.16.0
SQLAlchemy==1.4.26
telebot==0.0.4
telegram==0.0.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==3.0
urllib3==1.26.7
uuid==1.30
wcwidth==0.2.5
Werkzeug==2.0.2
